Partie 2/requirements.txt
google-generativeai==0.5.4
llama-index==0.12.5
llama-index-llms-gemini==0.4.1
llama-index-embeddings-gemini==0.3.0
streamlit==1.33.0
google-cloud-aiplatform==1.71.1